requirements.txt
pytesseract
numpy
opencv-python
pillow